+++ libs/qtdlib/messages/content/qtdmessagechataddmembers.h
@@ -17,16 +17,16 @@ class QTdMessageChatAddMembers : public QTdMessageContent
     Q_OBJECT
     Q_DISABLE_COPY(QTdMessageChatAddMembers)
     Q_PROPERTY(QObject *members READ members NOTIFY membersChanged)
-    Q_PROPERTY(qint32 firstMemberId READ firstMemberId NOTIFY membersChanged)
+    Q_PROPERTY(qint64 firstMemberId READ firstMemberId NOTIFY membersChanged)
 
 public:
     explicit QTdMessageChatAddMembers(QObject *parent = nullptr);
 
     QObject *members() const;
 
-    QList<qint32> memberUserIds() const;
-    qint32 firstMemberId() const;
-    void setSenderUserId(const qint32 senderUserId);
+    QList<qint64> memberUserIds() const;
+    qint64 firstMemberId() const;
+    void setSenderUserId(const qint64 senderUserId);
 
     void unmarshalJson(const QJsonObject &json);
 
@@ -36,8 +36,8 @@ signals:
 private:
     void updateTypeText();
     QScopedPointer<QTdUsersSortFilterModel> m_model;
-    QList<qint32> m_member_user_ids;
-    qint32 m_senderUserId;
+    QList<qint64> m_member_user_ids;
+    qint64 m_senderUserId;
 };
 
 #endif // QTDMESSAGECHATADDMEMBERS_H
